# specs/001-project-request-restaurant/contracts/api-contracts.yaml
openapi: 3.0.3
info:
  title: Restaurant Suite API
  description: Multi-role restaurant management and ordering system API
  version: 1.0.0
  contact:
    name: Restaurant Suite Team
servers:
  - url: https://api.restaurantsuite.example.com
    description: Production
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Creates a new user account (FR-001)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      description: Authenticates user and returns JWT tokens (FR-001)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Refreshes JWT access token using refresh token (FR-005)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token

  # ============ MENU ============
  /api/menu:
    get:
      tags: [Menu]
      summary: Get menu items
      description: Returns all menu items for the restaurant, optionally filtered (FR-022)
      parameters:
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
        - in: query
          name: available
          schema:
            type: boolean
      responses:
        '200':
          description: Menu items retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemDto'
  /api/admin/menu:
    post:
      tags: [Menu, Admin]
      summary: Create menu item
      description: Creates a new menu item (FR-016)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMenuItemRequest'
      responses:
        '201':
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/menu/{id}:
    put:
      tags: [Menu, Admin]
      summary: Update menu item
      description: Updates an existing menu item (FR-016)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMenuItemRequest'
      responses:
        '200':
          description: Menu item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemDto'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Menu, Admin]
      summary: Delete menu item
      description: Soft-deletes a menu item (FR-016)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Menu item deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ ORDERS ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, displayName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 10
        displayName:
          type: string
          minLength: 2
          maxLength: 100
        phone:
          type: string
          maxLength: 20
        role:
          type: string
          enum: [Admin, Waiter, Chef, Guest]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/UserDto'
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        role:
          type: string

    # ===== Menu =====
    CreateMenuItemRequest:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        description:
          type: string
          maxLength: 500
        price:
          type: number
          format: decimal
          minimum: 0
        sku:
          type: string
          maxLength: 50
        taxRate:
          type: number
          format: decimal
          minimum: 0
          maximum: 1
        isAvailable:
          type: boolean
          default: true
        imagePath:
          type: string
    UpdateMenuItemRequest:
      allOf:
        - $ref: '#/components/schemas/CreateMenuItemRequest'
    MenuItemDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        sku:
          type: string
        taxRate:
          type: number
          format: decimal
        isAvailable:
          type: boolean
        imagePath:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemOptionDto'
    MenuItemOptionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [Single, Multi]
        required:
          type: boolean
        modifiers:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemModifierDto'
    MenuItemModifierDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        priceDelta:
          type: number
          format: decimal

    # ===== Orders =====
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object